requirements.txt
diagrams==0.23.4
graphviz==0.20.1
Pillow==10.0.1
pyyaml==6.0.1
